replaced Long by long
git-svn-id: file:///home/svn/shoola-omero/trunk@3263 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/treeviewer/editors/EditorModel.java
@@ -346,11 +346,11 @@ class EditorModel
      */
     AnnotationData getAnnotationData()
     {
-        Long id = getUserDetails().getId();
+        long id = getUserDetails().getId();
         if (hierarchyObject == null) return null;
         else if ((hierarchyObject instanceof ImageData) || 
                 (hierarchyObject instanceof DatasetData))
-            return getLastAnnotation(getAnnotations(id.longValue()));
+            return getLastAnnotation(getAnnotations(id));
         return null;
     }
